+++ app/domain/entity/crp/body_crp.go
@@ -1,0 +1,9 @@
+package crp
+
+type ReqBody struct {
+	Data string `json:"data"`
+}
+
+type RspBody struct {
+	Data string `json:"data"`
+}