+++ app/infra/files/wisdom_infra.go
@@ -9,13 +9,13 @@ import (
 )
 
 // ParseJsonWisdom 从json解析wisdom
-func ParseJsonWisdom(filename string) (*entity.WisdomList, error) {
+func ParseJsonWisdom(filename string) (*entity.WisdomFileData, error) {
 	data, err := os.ReadFile(filename)
 	if err != nil {
 		return nil, errors.Wrap(err, "read file `wisdomHandler.json` got err")
 	}
 
-	var ws entity.WisdomList
+	var ws entity.WisdomFileData
 	err = json.Unmarshal(data, &ws)
 	if err != nil {
 		return nil, errors.Wrap(err, "unmarshal wisdomHandler json data got err")