+++ deployment/tools/parse_into_db.go
@@ -1,0 +1,13 @@
+package tools
+
+import (
+	"testing"
+)
+
+func TestName(t *testing.T) {
+	// json文件解析
+
+	// 初始化db
+
+	// db.BatchInsert
+}