+++ ChangeLog.md
@@ -1,6 +1,15 @@
 # ChangeLog
 
+## 2025/1/15
+
+1. Web日志、DB日志规范和重构
+2. 重构`GetOneWisdom`、`SaveWisdom`接口，基于DB来存储和查询金句
+3. 新增`ctx.GetHTTPReqEntry()`、`crp.xxReq`、`crp.xxRsp`来处理请求和响应参数
+4. 新增`tool`工具模块
+5. 解决了一些Bug
+
 ## 2025/1/2
+
 1. 重构了wisdom的框架处理流程，使之更切合WebHandler模式
 
 ## 2024/12/24